server/scripts/requirements.txt
rembg[cpu]
Pillow
numpy
